feat: apply common event page layout and confirm router name
