+++ src/app/channel/[channelName]/event/[eventId]/dashboard/page.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { NextPage } from 'next';
+
+import { Text } from '@/components/ui/text';
+
+const EventDashboardPage: NextPage = () => {
+  return <Text>Dashboard Page</Text>;
+};
+
+export default EventDashboardPage;
