+++ src/app/channel/[channelName]/event/[eventId]/event-setting/page.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { NextPage } from 'next';
+
+import { Text } from '@/components/ui/text';
+
+const EventSettingPage: NextPage = () => {
+  return <Text>Event Setting Page</Text>;
+};
+
+export default EventSettingPage;
