+++ src/app/channel/[channelName]/event/[eventId]/reservation/page.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { NextPage } from 'next';
+
+import { Text } from '@/components/ui/text';
+
+const EventReservationPage: NextPage = () => {
+  return <Text>Reservation Setting Page</Text>;
+};
+
+export default EventReservationPage;
