+++ src/app/channel/[channelName]/event/[eventId]/survey-setting/page.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { NextPage } from 'next';
+
+import { Text } from '@/components/ui/text';
+
+const EventSurveySettingPage: NextPage = () => {
+  return <Text>Survey Setting Page</Text>;
+};
+
+export default EventSurveySettingPage;
