+++ src/app/channel/[channelName]/event/[eventId]/time-setting/page.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { NextPage } from 'next';
+
+import { Text } from '@/components/ui/text';
+
+const EventTimeSettingPage: NextPage = () => {
+  return <Text>Time Setting Page</Text>;
+};
+
+export default EventTimeSettingPage;
