+++ src/components/design-token/color.ts
@@ -4,6 +4,7 @@ export const colorMap = {
   '#A1A1AA': '#A1A1AA',
   'gray/500': '#71717A',
   'gray/600': '#52525B',
+  '#6945F4': '#6945F4',
   none: 'none',
   currentColor: 'currentColor',
 } as const;
