+++ src/components/layout/index.ts
@@ -1,9 +1,7 @@
 import * as Header from './app-layout-header';
-import * as PageLayout from './app-layout-pagelayout';
 import * as SideMenu from './app-layout-sidemenu';
 
 export const AppLayout = {
   Header,
   SideMenu,
-  PageLayout,
 };
