+++ src/lib/typescript.ts
@@ -1,1 +1,3 @@
 export type ValueOf<T> = T[keyof T];
+
+export type NonEmptyArray<T> = [T, ...T[]];
